add password functionality I copied a function from the fernet docs on how to derive a key from a passcode
it uses randomly generated salt to create the key. We will be sending the salt along
with the locked up file. The user must have the passcode in order to derive the key
from the salt. This means that all security of this file is dependant upon the user
keeping this password secret.

Encrypt() -> generates random salt, saves it in salt.key, locks up an input file using
that salt, along with the user inputted passcode.

Decrypt() -> reads the salt.key file, asks for a passcode. It uses both of those in
order to derive a key, and decrypt the file. If the password is wrong, it will still
perform the action, you just won't get anything meaningful outputted.